Rename files to match commands
