+++ src/bin/cli.rs
@@ -3,9 +3,9 @@ use documents::Config;
 use documents::commands::export_fragments::{ExportFragmentsArgs, ExportFragmentsCommand};
 use documents::commands::health_checks::{HealthArgs, run as health_check};
 use documents::commands::list_all::ListAllCommand;
-use documents::commands::process_org::{ProcessOrgArgs, ProcessOrgCommand};
+use documents::commands::process_organization::{ProcessOrgArgs, ProcessOrgCommand};
 use documents::commands::process_repository::{ProcessRepositoryArgs, ProcessRepositoryCommand};
-use documents::commands::scan_org::{ScanOrgArgs, ScanOrgCommand};
+use documents::commands::scan_organization::{ScanOrgArgs, ScanOrgCommand};
 use documents::commands::serve_webhook::{ServeWebhookArgs, ServeWebhookCommand};
 use documents::commands::validate_repo_config::{ValidateConfigArgs, ValidateConfigCommand};
 use documents::github::{Client, GitHubClient};
