+++ src/commands/mod.rs
@@ -1,8 +1,8 @@
 pub mod export_fragments;
 pub mod health_checks;
 pub mod list_all;
-pub mod process_org;
+pub mod process_organization;
 pub mod process_repository;
-pub mod scan_org;
+pub mod scan_organization;
 pub mod serve_webhook;
 pub mod validate_repo_config;
